requirements.txt
Flask==3.0.2
Pillow==10.1.0
Pillow==10.2.0
rembg==2.0.56
